Help GUI + Answer Box Fix
The answer box was a simple fix, I simply made the answer variable a StringVar variable. I got rid of the "Pick an Answer" option as it was unnecessary.

I added in the help menu and added instructions based on how I think the quiz will work. It's possible the instructions will change later.